fix(leaks + cast): issue 8
Fix of compilation time error due to implicit cast.
Fix leaks due to modif in path_concat.c that should not
have been made.

+++ src/job/builtins/builtin_set.c
@@ -39,7 +39,7 @@ static void	display_shvar(void)
 			inside = ptr->next_content;
 			while (inside)
 			{
-				ft_printf("[%llu]=\"%s\" ", inside->index, inside->value);
+				ft_printf("[%llu]=\"%s\" ", (long long unsigned int)inside->index, inside->value);
 				inside = inside->next_content;
 			}
 			ft_printf(")\n");
